backend/requirements.txt
gunicorn==20.1.0
Flask==3.1.0
Flask-Cors==5.0.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.15.2
matplotlib==3.10.0
seaborn==0.13.2
tensorflow==2.19.0
requests==2.32.3
opencv-python==4.9.0.80
gunicorn==20.1.0
python-dotenv==1.0.0
